* commit '5ea2e1bbdd7d29a0ab43aa44a3b99a3678484e8c':
  #FREECB2020-43 #time 2h stworzenie widoku usuwania kategorii, aktualizacja widokow edycji i szczegolow
  poprawienie polskich znaków w stronach usera
  #FREECB2020-43 #time 1h walidowanie usuwania oferty, zmiana kategorii na inne, sprawdzanie czy unikalna nazwa przy dodawaniu

+++ app/Controllers/CategoryController.cs
@@ -10,7 +10,7 @@ using Microsoft.AspNetCore.Authorization;
 
 namespace freelancerzy.Controllers
 {
-    
+
     [Authorize(AuthenticationSchemes = "CookieAuthentication")]
     [Authorize(Roles = "administrator")]
     public class CategoryController : Controller
@@ -61,6 +61,11 @@ namespace freelancerzy.Controllers
         {
             if (ModelState.IsValid)
             {
+                if (_context.Category.First(c => c.CategoryName == category.CategoryName) != null)
+                {
+                    ViewBag.Message = "Nazwa musi być unikalna";
+                    return View(category);
+                }
                 _context.Add(category);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
@@ -143,6 +148,23 @@ namespace freelancerzy.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var category = await _context.Category.FindAsync(id);
+            if (category.CategoryName.Equals("Inne"))
+            {
+                ViewBag.Message = "Nie można usunąć tej kategorii!";
+                return View(category);
+            }
+
+            List<Offer> offers = await _context.Offer.Where(o => o.Category.Categoryid == category.Categoryid).ToListAsync();
+
+            Category defaultCategory = await _context.Category.FirstAsync(c => c.CategoryName == "Inne"); //TODO: gdzieś ustawić tą nazwę w configu
+
+            foreach (Offer o in offers)
+            {
+                o.Category = defaultCategory;
+                o.CategoryId = defaultCategory.Categoryid;
+                _context.Update(o);
+            }
+
             _context.Category.Remove(category);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
